pluginmanager: Move non-gui stuff to new plugin_comm
